Add support for specifying a parent column
For #140, this update allows a user to chose one of the existing columns that is a URI or URI Template type column as the parent for the property generated for another column. This is only a partial solution to #140 as it does not allow a user to group together a collection of columns (e.g. an address) without there being some existing column that can act as the parent for the group. To address this issue we need to also add support for creating virtual columns in the editor.