Move serilog configuration to logsettings.json
Updated startup so that Serilog configuration is read from a settings file rather than hardcoded.
Added logsettings.json files for web and worker roles with a basic configuration based on the original hard-coded configuration.
Changed the log index pattern to a monthly rotating name instead of a daily rotating name.