Add support for creating virtual columns
On the advanced view a virtual column can be added to the table schema. This must have a value which can be either a fixed URI or an URI template. Virtual columns can be used as the parent of one or more other columns, so in this way you can create a virtual column for the measurement and then assign values from other columns to be its children.
This doesn't provide a way to insert fixed literal metadata for a measurement though. Not sure if CSVW allows that, so if we do that it may need to be a non-standard extension to the CSVW spec.