Add support for measure columns (#211)
* Add support for creating virtual columns

On the advanced view a virtual column can be added to the table schema. This must have a value which can be either a fixed URI or an URI template. Virtual columns can be used as the parent of one or more other columns, so in this way you can create a virtual column for the measurement and then assign values from other columns to be its children.
This doesn't provide a way to insert fixed literal metadata for a measurement though. Not sure if CSVW allows that, so if we do that it may need to be a non-standard extension to the CSVW spec.

* Allow virtual columns to have a fixed literal value as well as URI / URI Template values.

* Fix for portal display when the search button property in user settings is null

* Refactor UI to separate CSVW model from the UI view model

* Refactor AdvancedRow

Reuse the standard row fields in the "value" section of the measure column.
Fix PrefixedUriInput reactivity problems so that it now accepts an external change to the model

* Only notify change to expanded uri on blur

* Add support for creating new facets for measure columns

* Add support for removing a facet from a measure column

* Bug fixes for handling measure columns in the ui

Co-authored-by: Kal <kal@techquila.com>